refactor: remove Board class
The Board class was removed, and all its functionality was moved to the GraphicalUI class. This simplifies the architecture of the application by establishing the GraphicalUI as the central controlling class for all user interface elements and operations, leading to clearer code organization. Menu, Entry, MenuHandler, and Label classes were updated to receive a new 'parent' parameter, which is the GraphicalUI instance. This change allows for easy access to the main application methods and attributes, thereby promoting loose coupling between these classes and enhancing code maintainability. Board class methods and attributes have been integrated into GraphicalUI.
