+++ src/interfaces/graphicalui/__init__.py
@@ -1,4 +1,3 @@
-from .board import Board
 from .button import Button
 from .entry import Entry
 from .label import Label
