+++ src/interfaces/graphicalui/button.py
@@ -3,44 +3,20 @@ from typing import Callable
 
 
 class Button:
-    """
-    Represents a button on a game board.
-
-    Attributes:
-        INITIAL_HORIZONTAL_POSITION (int): The initial horizontal position for buttons.
-        INITIAL_VERTICAL_POSITION (int): The initial vertical position for buttons.
-        DOUBLE_SWAP_HORIZONTAL_POSITION (int): The horizontal position for buttons when double swapping.
-        BUTTON_HEIGHT (int): The height of the button.
-    """
-
     INITIAL_HORIZONTAL_POSITION = 100
     INITIAL_VERTICAL_POSITION = 210
     DOUBLE_SWAP_HORIZONTAL_POSITION = 67
     BUTTON_HEIGHT = 30
 
     def __init__(self, parent, double_swap: bool, swap_count: int, command: Callable):
-        """
-        Initializes a Button instance.
-
-        Args:
-            parent: The parent widget that contains the button.
-            double_swap (bool): Whether it's a double swap button.
-            swap_count (int): The number of swaps associated with the button.
-            command (Callable): The function to execute when the button is clicked.
-        """
-        self.parent = parent
+        self.parent = parent.window
         self.double_swap = double_swap
         self.swap_count = swap_count
         self.command = command
         self.button = self.initialize()
 
     def initialize(self):
-        """
-        Initializes the button, sets its style, and positions it on the screen.
-        Returns:
-            ttk.Button: The initialized button.
-        """
-        master = self.parent.window
+        master = self.parent
         text = self.get_label()
         command = self.command
 
@@ -52,12 +28,6 @@ class Button:
         return button
 
     def get_label(self):
-        """
-        Generates the label text for the button based on the swap count.
-
-        Returns:
-            str: The label text.
-        """
         return (
             f"{self.swap_count} Swap"
             if self.swap_count == 1
@@ -65,12 +35,6 @@ class Button:
         )
 
     def set_position(self, button):
-        """
-        Sets the position of the button on the screen.
-
-        Args:
-            button (ttk.Button): The button to position.
-        """
         width, height = self.calculate_size()
         horizontal_position, vertical_position = self.calculate_position()
 
@@ -79,12 +43,6 @@ class Button:
         )
 
     def horizontal_position(self):
-        """
-        Calculates the horizontal position based on double swap and swap count.
-
-        Returns:
-            int: The horizontal position.
-        """
         return (
             self.DOUBLE_SWAP_HORIZONTAL_POSITION * self.swap_count
             if self.double_swap
@@ -92,12 +50,6 @@ class Button:
         )
 
     def calculate_width(self):
-        """
-        Calculates the button width based on double swap.
-
-        Returns:
-            int: The button width.
-        """
         return (
             self.DOUBLE_SWAP_HORIZONTAL_POSITION
             if self.double_swap
@@ -105,36 +57,18 @@ class Button:
         )
 
     def calculate_size(self):
-        """
-        Calculates the button size (width and height).
-
-        Returns:
-            Tuple[int, int]: The button width and height.
-        """
         width = self.calculate_width()
         height = self.BUTTON_HEIGHT
 
         return width, height
 
     def calculate_padding(self):
-        """
-        Calculates the horizontal and vertical padding for button positioning.
-
-        Returns:
-            Tuple[int, int]: The horizontal and vertical padding.
-        """
         horizontal_padding = 25
         vertical_padding = 25
 
         return horizontal_padding, vertical_padding
 
     def calculate_position(self):
-        """
-        Calculates the position of the button on the screen.
-
-        Returns:
-            Tuple[int, int]: The horizontal and vertical position.
-        """
         horizontal_padding, vertical_padding = self.calculate_padding()
 
         horizontal_position = horizontal_padding + self.horizontal_position()
@@ -144,9 +78,6 @@ class Button:
 
     @staticmethod
     def configure_style():
-        """
-        Configures the style of the ttk button.
-        """
         style = ttk.Style()
         style.configure(
             style="TButton",
